+++ database/migrations/2023_10_18_230742_create_plans_table.php
@@ -19,7 +19,7 @@ class CreatePlansTable extends Migration
             $table->string('description')->nullable();
             $table->string('slug');
             $table->string('reference');
-            $table->double('price', 10, 2);
+            $table->integer('price');
             $table->timestamps();
         });
     }
